refactor: Move TxDetails in TxForm folder since only used by that component
